+++ NEWS.md
@@ -5,6 +5,7 @@
 * `iMRCSingle()`:  Added.  Was `mrc1()` internal function inside of `mrClosed()` environment.
 * `lrt()`: Modified.  Slight change to row labels in output table.
 * `mrClosed()`: Modified.  Moved all internal functions outside of `mrClosed()` environment (see `iMRCMultiple()` and `iMRCSingle()`).  Added more catches for argument problems (required setting `n=`, `m=`, `M=` and `R=` to `NULL`).  
+* `plot.mrClosed()`: Modified.  Removed setting of `par()`.  Changed from using `lowess()` to using `loess()` and set better default values.  Added descriptive text to help file.
 
 # FSA 0.4.16 Jul14
 * `BluegillLM`: Modified.  Added a seealso.